% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_2_sim.R
\name{cal_2_sim}
\alias{cal_2_sim}
\title{The cal_2_sim() function to calibrate (poststratify or rake) an outcome onto simulted poststrata totals}
\usage{
cal_2_sim(des_2be_cal, df_or_list_sim, form_outcome, form_poststrat, type_cal)
}
\arguments{
\item{des_2be_cal}{is a ?survey.design object created by ?survey::svydesign}

\item{df_or_list_sim}{1 simulation of single data.frame (for joint) or single list of data.frames (for marginals)}

\item{form_outcome}{formula for the outcome}

\item{form_poststrat}{single formula of joint poststrata or list of formulas for marginal poststrata}

\item{type_cal}{either mcsp or mcsr (MCS by poststratification or raking)}
}
\value{
a list of data.frames with c("total","SE") as columns
}
\description{
The cal_2_sim() function to calibrate (poststratify or rake) an outcome onto simulted poststrata totals
}
\examples{

}
