% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_est.R
\name{combine_est}
\alias{combine_est}
\title{The combine_est() function to combine intermediate calibrated estimates}
\usage{
combine_est(list_cal)
}
\arguments{
\item{list_cal}{list containing intermediate estimates that have been calibrated by \code{\link[MultCal2Sim]{mult_cal}}}
}
\value{
a 1 row data.frame of the MCS estimate and its standard error
}
\description{
The combine_est() function to combine intermediate calibrated estimates
}
\examples{

}
