% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_tot_ref.R
\name{est_tot_from_des_1marg}
\alias{est_tot_from_des_1marg}
\title{helper function to estimate ONE marginal total from a survey design object}
\usage{
est_tot_from_des_1marg(form_1_term_only, design_refer)
}
\arguments{
\item{form_1_term_only}{a formula with only 1 term on the right hand side representing a marginal total}

\item{design_refer}{a ?survey::design object that contains the survey data used to estimate the marginal total}
}
\value{
result is 'tidy' data frame
}
\description{
helper function to estimate ONE marginal total from a survey design object
}
\examples{
dstrat <- svydesign(id=~1,strata=~stype, weights=~pw, data=apistrat, fpc=~fpc)
est_tot_from_des_1marg(~sch.wide,design_refer=dstrat)
est_tot_from_des_1marg(~both,design_refer=dstrat)
lapply(list(~sch.wide,~both),est_tot_from_des_1marg,design_refer=dstrat)

}
