% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_tot_ref.R
\name{est_tot_from_des_joint}
\alias{est_tot_from_des_joint}
\title{helper function to estimate joint totals from survey design object}
\usage{
est_tot_from_des_joint(form_additive, design_refer)
}
\arguments{
\item{form_additive}{a formula with only right hand side additive terms representing joint totals}

\item{design_refer}{a survey.design object that contains the survey data used to estimate the joint total(s) \code{\link[survey]{svydesign}}}
}
\value{
result is 'tidy' data frame, row is unique strata combo
}
\description{
helper function to estimate joint totals from survey design object
}
\examples{

dstrat <- svydesign(id=~1,strata=~stype, weights=~pw, data=apistrat, fpc=~fpc)
est_tot_from_des_joint(~sch.wide+bothdesign_refer=dstrat)
}
