% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcs.R
\name{mcs}
\alias{mcs}
\title{The 'mcs()' helper function combines the 3-steps: simulate, calibrate, combine}
\usage{
mcs(des_2be_cal, df_or_list_est_tot, form_outcome, form_poststrat, type_cal,
  num_sim = 1, parallel = FALSE, num_core = 1, lgl_rej_neg_sim = TRUE)
}
\arguments{
\item{des_2be_cal}{a survey design object that is to be calibrated see \code{\link[survey]{design}}}

\item{df_or_list_est_tot}{a single data.frame of estimated joint totals or list of data.frames of estimated marginal totals (and respective standard errors) see \code{\link[MultCal2Sim]{sim_tot_from_est}}}

\item{form_outcome}{a formula of the outcome, whose total we want an improved MCS estimate of see \code{\link[MultCal2Sim]{cal_2_sim}}}

\item{form_poststrat}{single formula of joint poststrata or list of formulas for marginal poststrata see \code{\link[MultCal2Sim]{cal_2_sim}}}

\item{type_cal}{a character string for 'mcsp' or 'mcsr'}

\item{num_sim}{the number of simulations to calibrate onto}

\item{parallel}{a logical whether to run in parallel using \code{\link[parallel]{parLapply}}}

\item{num_core}{the number of cores for the parallel job}

\item{lgl_rej_neg_sim}{a logical whether to reject negative simulations}
}
\value{
a 1 row data.frame of the MCS estimate and its standard error
}
\description{
The 'mcs()' helper function combines the 3-steps: simulate, calibrate, combine
}
\examples{

}
