% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_tot_from_est.R
\name{sim_tot_from_est}
\alias{sim_tot_from_est}
\title{The sim_tot_from_est() function to simulate totals from pre-existing estimates}
\usage{
sim_tot_from_est(df_or_list_est_tot, type_strata, lgl_rej_neg_sim = TRUE)
}
\arguments{
\item{df_or_list_est_tot}{a single data.frame of estimated joint totals or list of data.frames of estimated marginal totals (and respective standard errors)}

\item{type_strata}{specify if joint (requires a dataframe) or marginal requires a list}

\item{lgl_rej_neg_sim}{logical to reject negative simulations of totals}
}
\value{
1 simulation of either joint totals in a data.frame or marginal totals in a list of data.frames
}
\description{
The sim_tot_from_est() function to simulate totals from pre-existing estimates
}
\examples{

}
